Moved Scale to the post-game page
