+++ Scouting/src/main/java/ca/team2706/scouting/mcmergemanager/datamodels/PostGameObject.java
@@ -1,6 +1,8 @@
 package ca.team2706.scouting.mcmergemanager.datamodels;
 
 import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Created by MCSoftware on 2016-01-23.
@@ -9,14 +11,16 @@ public class PostGameObject implements Serializable {
     public String notes;
     public boolean challenged;
     public int timeDead;
+    public List<ScalingTime> scalingTower;
 
     public PostGameObject() {
-
+        scalingTower = new ArrayList<>();
     }
 
-    public PostGameObject(String notes, boolean challenged, int timeDead) {
+    public PostGameObject(String notes, boolean challenged, int timeDead, List<ScalingTime> scalingTower) {
         this.notes = notes;
         this.challenged = challenged;
         this.timeDead = timeDead;
+        this.scalingTower = scalingTower;
     }
 }
