+++ Scouting/src/main/java/ca/team2706/scouting/mcmergemanager/datamodels/TeleopScoutingObject.java
@@ -28,21 +28,18 @@ public class TeleopScoutingObject implements Serializable {
     public List<Integer> defensesBreached;
     public double timeDefending;
     public List<BallPickup> ballsPickedUp;
-    public List<ScalingTime> scalingTower;
 
     public TeleopScoutingObject() {
         ballsShot = new ArrayList<>();
         defensesBreached = new ArrayList<>();
         ballsPickedUp = new ArrayList<>();
-        scalingTower = new ArrayList<>();
     }
 
-    public TeleopScoutingObject(List<BallShot> ballsShot,List<Integer> defensesBreached, double timeDefending,List<BallPickup> ballsPickedUp,List<ScalingTime> scalingTower) {
+    public TeleopScoutingObject(List<BallShot> ballsShot,List<Integer> defensesBreached, double timeDefending,List<BallPickup> ballsPickedUp) {
         this.ballsShot = ballsShot;
         this.defensesBreached = defensesBreached;
         this.timeDefending = timeDefending;
         this.ballsPickedUp = ballsPickedUp;
-        this.scalingTower = scalingTower;
     }
 
     public static String getDefenseName(int i) {
